% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO.R
\name{CO_flux}
\alias{CO_flux}
\title{CO_flux - Calculates annual deposition of CO on vegetation}
\usage{
CO_flux(x)
}
\arguments{
\item{x}{A data frame containing hourly data of CO concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Annual value of CO deposited (g m-2 yr-1)
}
\description{
CO_flux - Calculates annual deposition of CO on vegetation
}
\examples{

data(Bizkaia_data)
CO_flux(x = Bizkaia_data)

}
