% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO_h.R
\name{CO_flux_h}
\alias{CO_flux_h}
\title{CO_flux_h - Calculates hourly deposition of CO on vegetation}
\usage{
CO_flux_h(x)
}
\arguments{
\item{x}{A data frame containing hourly data of CO concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of CO deposited (g m-2 yr-1)
}
\description{
CO_flux_h - Calculates hourly deposition of CO on vegetation
}
\examples{

data(Bizkaia_data)
CO_flux_h(x = Bizkaia_data)

}
