% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conc_CO.R
\name{Conc_CO}
\alias{Conc_CO}
\title{Conc_CO - Extracts data of hourly concentration of CO}
\usage{
Conc_CO(x)
}
\arguments{
\item{x}{A data frame containing hourly data of CO concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of concentration of CO (micrograms m-3)
}
\description{
Conc_CO - Extracts data of hourly concentration of CO
}
\examples{

data(Bizkaia_data)
Conc_CO(x = Bizkaia_data)
}
