% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conc_NO2.R
\name{Conc_NO2}
\alias{Conc_NO2}
\title{Conc_NO2 - Extracts data of hourly concentration of NO2}
\usage{
Conc_NO2(x)
}
\arguments{
\item{x}{A data frame containing hourly data of NO2 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of concentration of NO2 (micrograms m-3)
}
\description{
Conc_NO2 - Extracts data of hourly concentration of NO2
}
\examples{

data(Bizkaia_data)
Conc_NO2(x = Bizkaia_data)
}
