% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conc_O3.R
\name{Conc_O3}
\alias{Conc_O3}
\title{Conc_O3 - Extracts data of hourly concentration of O3}
\usage{
Conc_O3(x)
}
\arguments{
\item{x}{A data frame containing hourly data of O3 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of concentration of O3 (micrograms m-3)
}
\description{
Conc_O3 - Extracts data of hourly concentration of O3
}
\examples{

data(Bizkaia_data)
Conc_O3(x = Bizkaia_data)
}
