% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Conc_PM10.R
\name{Conc_PM10}
\alias{Conc_PM10}
\title{Conc_PM10 - Extracts data of hourly concentration of PM10}
\usage{
Conc_PM10(x)
}
\arguments{
\item{x}{A data frame containing hourly data of PM10 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of concentration of PM10 (micrograms m-3)
}
\description{
Conc_PM10 - Extracts data of hourly concentration of PM10
}
\examples{

data(Bizkaia_data)
Conc_PM10(x = Bizkaia_data)
}
