% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Daylight.R
\name{Daylight}
\alias{Daylight}
\title{LAI_evergreen - Generates hourly data of leaf and bark area index for evergreen trees in a specific year}
\usage{
Daylight(shortest_day_sunrise, shortest_day_sunset, longest_day_sunset0, Year)
}
\arguments{
\item{shortest_day_sunrise}{Sunrise time in the shortest day in the Northern Hemisphere (December 21) using decimals for minutes (e.g. 8.4)}

\item{shortest_day_sunset}{Sunset time in the shortest day in the Northern Hemisphere (December 21) using decimals for minutes (e.g. 17.8)}

\item{longest_day_sunset0}{Sunset time in the longest day in the Northern Hemisphere (June 21) using decimals for minutes (e.g. 21.9)}

\item{Year}{Year to generate hourly daylight data (e.g. 2015)}
}
\value{
A dataframe with LAI and BAI hourly values is generated
}
\description{
LAI_evergreen - Generates hourly data of leaf and bark area index for evergreen trees in a specific year
}
\examples{

Daylight (shortest_day_sunrise  = 8.4,
shortest_day_sunset = 17.8,
longest_day_sunset0 = 21.9,
Year = 2016)

}
