% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dep_NO2.R
\name{Dep_NO2}
\alias{Dep_NO2}
\title{Dep_NO2 - Calculates hourly deposition of NO2 on vegetation}
\usage{
Dep_NO2(x)
}
\arguments{
\item{x}{A data frame containing hourly data of NO2 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of deposition of NO2 on vegetation (g m-2 h-1)
}
\description{
Dep_NO2 - Calculates hourly deposition of NO2 on vegetation
}
\examples{

data(Bizkaia_data)
Dep_NO2(x = Bizkaia_data)
}
