% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dep_PM10.R
\name{Dep_PM10}
\alias{Dep_PM10}
\title{Dep_PM10 - Calculates hourly deposition of PM10 on vegetation}
\usage{
Dep_PM10(x, z_0 = 1)
}
\arguments{
\item{x}{A data frame containing hourly data of PM10 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}

\item{z_0}{Roughness length value (m)}
}
\value{
Hourly data of deposition of PM10 on vegetation (g m-2 h-1)
}
\description{
Dep_PM10 - Calculates hourly deposition of PM10 on vegetation
}
\examples{

data(Bizkaia_data)
Dep_PM10(x = Bizkaia_data, z_0 = 1)
}
