% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Dep_vel_CO.R
\name{Dep_vel_CO}
\alias{Dep_vel_CO}
\title{Dep_vel_CO - Calculates hourly deposition velocity for CO}
\usage{
Dep_vel_CO(x)
}
\arguments{
\item{x}{A data frame containing hourly data of weather variables (e.g. Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight))}
}
\value{
Hourly data of deposition velocity for CO (m s-1)
}
\description{
Dep_vel_CO - Calculates hourly deposition velocity for CO
}
\examples{

data(Bizkaia_data)
Dep_vel_CO(x = Bizkaia_data)
}
