% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LAI_evergreen.R
\name{LAI_evergreen}
\alias{LAI_evergreen}
\title{LAI_evergreen - Generates hourly data of leaf and bark area index for evergreen trees in a specific year}
\usage{
LAI_evergreen(Year, LAI_value, BAI_value)
}
\arguments{
\item{Year}{Year to generate leaf and bark area index (e.g. 2015)}

\item{LAI_value}{Mean value of leaf area index (e.g. 1.3)}

\item{BAI_value}{Bark area index value (e.g. 0.1)}
}
\value{
A dataframe with LAI and BAI hourly values is generated
}
\description{
LAI_evergreen - Generates hourly data of leaf and bark area index for evergreen trees in a specific year
}
\examples{

LAI_evergreen(Year = 2016,
BAI_value = 0.1,
LAI_value = 1.3)

}
