% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/NO2.R
\name{NO2_flux}
\alias{NO2_flux}
\title{NO2_flux - Calculates annual deposition of NO2 on vegetation}
\usage{
NO2_flux(x)
}
\arguments{
\item{x}{A data frame containing hourly data of NO2 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Annual value of NO2 deposited (g m-2 yr-1)
}
\description{
NO2_flux - Calculates annual deposition of NO2 on vegetation
}
\examples{

data(Bizkaia_data)
NO2_flux(x = Bizkaia_data)

}
