% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O3.R
\name{O3_flux}
\alias{O3_flux}
\title{O3_flux - Calculates annual deposition of O3 on vegetation}
\usage{
O3_flux(x)
}
\arguments{
\item{x}{A data frame containing hourly data of O3 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Annual value of O3 deposited (g m-2 yr-1)
}
\description{
O3_flux - Calculates annual deposition of O3 on vegetation
}
\examples{

data(Bizkaia_data)
O3_flux(x = Bizkaia_data)

}
