% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/O3_h.R
\name{O3_flux_h}
\alias{O3_flux_h}
\title{O3_flux_h - Calculates hourly deposition of O3 on vegetation}
\usage{
O3_flux_h(x)
}
\arguments{
\item{x}{A data frame containing hourly data of O3 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of O3 deposited (g m-2 yr-1)
}
\description{
O3_flux_h - Calculates hourly deposition of O3 on vegetation
}
\examples{

data(Bizkaia_data)
O3_flux_h(x = Bizkaia_data)

}
