% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PM10_h.R
\name{PM10_flux_h}
\alias{PM10_flux_h}
\title{PM10_flux_h - Calculates hourly deposition of PM10 on vegetation}
\usage{
PM10_flux_h(x)
}
\arguments{
\item{x}{A data frame containing hourly data of PM10 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of PM10 deposited (g m-2 yr-1)
}
\description{
PM10_flux_h - Calculates hourly deposition of PM10 on vegetation
}
\examples{

data(Bizkaia_data)
PM10_flux_h(x = Bizkaia_data)

}
