% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_meso_SO2.R
\name{Res_meso_SO2}
\alias{Res_meso_SO2}
\title{Res_meso_SO2 - Calculates hourly mesophyll resistance for SO2}
\usage{
Res_meso_SO2(x)
}
\arguments{
\item{x}{A data frame containing hourly data of weather variables (e.g. Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight))}
}
\value{
Hourly data of mesophyll resistance for SO2 (s m-1)
}
\description{
Res_meso_SO2 - Calculates hourly mesophyll resistance for SO2
}
\examples{

data(Bizkaia_data)
Res_meso_SO2(x = Bizkaia_data)
}
