% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_soil.R
\name{Res_soil}
\alias{Res_soil}
\title{Res_soil - Calculates soil resistance on an hourly basis}
\usage{
Res_soil(x)
}
\arguments{
\item{x}{A data frame containing hourly data of weather and other variables (e.g. Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Hourly data of soil resistance (s m-1)
}
\description{
Res_soil - Calculates soil resistance on an hourly basis
}
\examples{

data(Bizkaia_data)
Res_soil(x = Bizkaia_data)
}
