% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_Tot_CO.R
\name{Res_Tot_CO}
\alias{Res_Tot_CO}
\title{Res_Tot_CO - Calculates hourly total resistance for CO}
\usage{
Res_Tot_CO(x)
}
\arguments{
\item{x}{A data frame containing hourly data of weather variables (e.g. Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight))}
}
\value{
Hourly data of total resistance for CO (s m-1)
}
\description{
Res_Tot_CO - Calculates hourly total resistance for CO
}
\examples{

data(Bizkaia_data)
Res_Tot_CO(x = Bizkaia_data)
}
