% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Res_Tot_O3.R
\name{Res_Tot_O3}
\alias{Res_Tot_O3}
\title{Res_Tot_O3 - Calculates hourly total resistance for O3}
\usage{
Res_Tot_O3(x, z_0 = 1)
}
\arguments{
\item{x}{A data frame containing hourly data of weather variables (e.g. Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight))}

\item{z_0}{Roughness length value (m)}
}
\value{
Hourly data of total resistance for O3 (s m-1)
}
\description{
Res_Tot_O3 - Calculates hourly total resistance for O3
}
\examples{

data(Bizkaia_data)
Res_Tot_O3(x = Bizkaia_data, z_0 = 1)
}
