% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SO2.R
\name{SO2_flux}
\alias{SO2_flux}
\title{SO2_flux - Calculates annual deposition of SO2 on vegetation}
\usage{
SO2_flux(x)
}
\arguments{
\item{x}{A data frame containing hourly data of SO2 concentration and other variables (Dates (e.g. 01/01/2016  00:00:00), Hum (\%), Pres (kPa), Precip (mm), Rad (W m-2), Temp (C), Wind (m s-1), Daylight (Night or Daylight), BAI, LAI)}
}
\value{
Annual value of SO2 deposited (g m-2 yr-1)
}
\description{
SO2_flux - Calculates annual deposition of SO2 on vegetation
}
\examples{

data(Bizkaia_data)
SO2_flux(x = Bizkaia_data)

}
