% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeDep.R
\docType{package}
\name{TreeDep}
\alias{TreeDep}
\alias{TreeDep-package}
\title{TreeDep Package}
\description{
The model estimates air pollution removal by dry deposition on trees. The pollutans are nitrogen dioxide, ozone, sulphur dioxide, carbon monoxide and particulate matter.
}
\author{
Silvestre Garcia de Jalon \email{s.garciadejalon@gmail.com}
}
