% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TreeDep.R
\docType{package}
\name{TreeDep}
\alias{TreeDep}
\alias{TreeDep-package}
\title{TreeDep Package}
\description{
The model estimates air pollution removal by dry deposition on trees. It also estimates aerodynamic resistance, boundary layer resistance, canopy resistance, stomatal resistance, cuticular resistance, mesophyll resistance, soil resistance, friction velocity and deposition velocity. It also allows plotting graphical results for a specific time period. The pollutants are nitrogen dioxide, ozone, sulphur dioxide, carbon monoxide and particulate matter.
}
\author{
Silvestre Garcia de Jalon \email{s.garciadejalon@gmail.com}
}
